refactor: environment module layout
